Freeplay Lane + Jump pack Int.
Lots have changed here, let's break it down:
The lane method of generation has began development for freeplay. This means most gameplay features will be separated per gamemode, and each will have its own function file. This function file will contain most of the code to run it's own gamemode.

Currently, the game is in a weird state where the freeplay mode uses some of the old code and sometimes uses new code, so do not remove any old code yet! I've also added switches to most old code to specifically use certain lines for when within a lane (this is done with execute if dimension). In some point in the future when the full rewrite to Freeplay.mcfunction is complete, all the old code can be erased. This is currently developed in a way to keep all old code and methods in tact. To reach the new Freeplay location, act as if you were summoning a teleporter to the original game but toss in a gold ingot instead of a diamond.

Jump packs are starting to be integrated. Currently, the default jump pack is always selected, in the future we want to player to choose this for themselves. Right now, this corresponds to the jump pack loaded in the editor. The game now generates these jumps instead of rogue jumps. It chooses a jump at random from the pack, places all of the markers needed for it's blocks (this includes all platforms and blockers currently, no pickups or other blocks). Then, the code just treats these as the usual blocks from before and works pretty much the same as before. Some new tags: ParkourGoal is the block the player must reach to complete a jump (this is on the marker). ParkourGenPos is the block the game will create the next jump at when the player reaches ParkourGoal (this is also on a marker). The tag ParkourPrevGoal is the block the player has just jumped to and completed the previous jump from (this is only used in line.mcfunction for the trails, no other definitions are used currently.)

Trails are currently temporarily implemented somewhat simply. Right now, they will simply draw a line from the previous goal block, through the current goal block, to the next goal block (ParkourGenPos tag). In the future, we want the player to define this path within their jumppacks that they create.

Particle effects currently only play when the player reaches their goal block (and that's when block increments.) Should we switch to a distance type of scoring? Should particles and effects play on every block?

Mirroring and randomness bias are both currently not implemented, that is on my to-do list for tomorrow.

Blockers currently summon yellow stained glass, which act the same as platforms currently. Since none of the effects or incrementing of scores happen on ParkourNextJump blocks, these blockers have no effect. Keep this in mind if these properties change in the future, and purposefully exclude the tag ParkourBlocker when detecting landing and such.

Decorations are one of the main things not implemented directly into Freeplay mode's function, if someone would like to get that in that would be really helpful.