Distance instead of Jumps
A change has been made to the score we store for the player. Instead of counting the number of jumps made, we are now counting the distance they travel from the start in the Z direction (Z - 12). This is preferred so that longer and larger jumps provide a better score for the player (and in the future, when time is also stored and scored, longer jumps will save time in theory.) In addition, Highscore is being constantly set whenever the player beats their best, and their distance is also being constantly set (in scoreboard Blocks).