Reintroducing Player Teams
In order for summoned slime entities that create outlines for blocks that are unable to get one to not collide with the player, Three teams have been reintroduced: ParkourPlayers for all players inside the game actively parkouring, Highscore for any player at the lobby, and infpar_editor for anyone inside the editor. Slimes in the editor will no longer collide with the player.