Block Dictionary + JSON Compiler
Block Dictionary is a new JSON file that stores all information about all of the blocks that you can find while parkouring or building packs. The goal is to make it so you only need to change one file to add new blocks, not 7 different places within 3 different files. Currently, only block displays use dictionary searches, but soon all references of blocks will use the dictionary search, I just gotta implement it.

Currently, the JSON will not be imported if you change it, because I need the new compiler to work to get there. It's almost done, the only problem is that the compiler deletes the imports when automatic mode is first used, and then it inserts two imports of the same data every time that the auto-compiler runs. As soon as this bug is figured out, then it will work. If you want to use the new compiler, modify build.bat and autobuild.bat to have "JSON" after the "MinecraftScript" declaration. That will swap to the jar that I've been working on, *all day.* Any help I could get on this compiler I would be forever greatful!