Hologram Dictionary Searches
New thing learned: If using data that is optional for blocks within a search, you must clear them from the search buffer before handling the next block, otherwise you get trailing data. This technically means other searches contain trailing data as well, but it isn't as much of a problem for some reason, depends on the Search's use case. Annoying, but I think it might be unavoidable as all data has to be merged into the search buffer so that it's all usable via macros and their one level of depth accessibility. Perhaps later it will become a bigger issue, or will cause headaches while debugging since it is a bit of a silent problem.