Beginning of Rotational Implementation
Blocks can now be defined as rotational, by setting the tag "rotational":true. Non-rotational blocks DO NOT need to set this value to false, and if you do so, they WILL be treated as rotational, since the code simply checks for the prescence of the tag inside of the block dictionary.

Displays inside of the editor now work to be properly rotated with the placed block. This is based on the item frame that is used to place the block. This currently means that it is possible to place a ladder on the tops or bottoms of other blocks.

TO DO: Saving, Loading, and the hologram need to be able to handle new rotational information found in Jump packs. Placement of blocks inside of the game also has no support for this just yet.