Rotational Blocks (Horizonal) Complete!
Any block that is defined as rotational inside of the block dictionary (like ladders) can now be used in any of their rotations within Infinite Parkour! I recommend reading previous commit notes as they provide more information on how this works exactly. This commit here adds a new tag to every marker called ip_rot_# where the # is replaced with the facing value that an item frame in that position would have (more info on that in last commit). A special check also is that these tags are mirrored east and westward in accordance to the #jump_mirror_math scoreboard, which is set to either 1 when a jump is normal, and -1 when a jump is mirrored.

In addition, a precaution has been added where all players will kill any items around themselves while playing in a 64 block area unless that item is tagged with ip_item. This could be useful later, and maybe temporary as I've heard using distance selectors isn't always the best option. This was needed as Ladders kept dropping as items since when their support block is broken, they would drop.