Ice, Soul Sand, Cobweb + Bug Fix
Added 3 new (and easy) blocks, that being Ice, Soul Sand, and Cobwebs! The last I'm still on the fence for, as it could make for a lot more boring gameplay rather than anything fun or interesting. Also, the Z-Fighting it has with its display in the editor is really bad and kind of unavoidable (it's a crisscross, you can't translate the thing in any direction to remove the z-fighting).

A bug was fixed that made non-rotational blocks place like they were rotational due to trailing data.