Basis of Palette System Introduction
So the idea for the next major update includes defining palettes for the different areas and zones, these will be stored in a dictionary that has been created. Right now, only the default palette has been introduced, and currently THIS PALETTE IS HARDCODED. What I did is include references to the default palette directly instead of a searching algorithm which will be needed in the future. We need to figure out a place to store the player's current palette, rather that is per player or per lobby or even per jump or jumppack. Probably the last option.