Separate Highscores Per Pack!
Highly requested feature that was actually easy to implement! This does mean ALL SCORES WILL BE RESET AFTER THIS VERSION FROM LAST VERSION, sorry.

TO DO: The cleanup function does not remove all the new scoreboards that are generated with this change. I opted for creating new scoreboards per pack so that others can place on this scoreboard against each other (for true multiplayer support in the future.)