5 Searches into 1 + Loot table Templates
To save on performance, I have cut down on dictionary references by storing this information on the markers for each block. Then, I give that data to the block displays after, from the markers. This allows for less loops to happen. Flaming_Thunder_ added some templates for loot tables as we plan to use those for implementations later.

Co-Authored-By: Flaming <180693730+Flaming-Thunder@users.noreply.github.com>
Co-Authored-By: Chilenderino <189549378+Chilenderino@users.noreply.github.com>
Co-Authored-By: OfekN_ <83116238+ofek2608@users.noreply.github.com>